+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Batch 18] - 2025-07-12
+
+### Added
+- `DB_RETRIES` and `DB_RETRY_DELAY_MS` environment variables for executor DB connection retries
+
 ## [Batch 17] - 2025-07-11
 
 ### Added